alphabot/requirements.txt
google-adk==0.1
httpx==0.28.1
pandas==2.2.3
